Allow parallelize the optimization of framework-res.apk
This change enhancement the optimization of cm ota package, allowing do
the optimization of framework-res.apk in a parallel way. The current
script optimize first the framework-res package and then the rest of
packages with xargs in parallel. The framework-res package can be
processed with xargs in parallel copying all the packages to a
temporally folder to apply xargs (because opticharged need that the
package is the current directory). This shortens the time of
optimization process, because while framework-res package is optimized,
the rest of processor cores can be optimizing other packages.